+++ packages/windie-sdk-js/cjs/index.js
@@ -29,6 +29,7 @@ __exportStar(require("./projections/conversationProjections.js"), exports);
 __exportStar(require("./runtime/conversationReducer.js"), exports);
 __exportStar(require("./runtime/conversationEventScope.js"), exports);
 __exportStar(require("./runtime/ConversationRuntime.js"), exports);
+__exportStar(require("./runtime/WindieConversationRuntime.js"), exports);
 __exportStar(require("./runtime/TraceRecorder.js"), exports);
 __exportStar(require("./runtime/TurnInputPipeline.js"), exports);
 __exportStar(require("./runtime/DefaultTurnResourceResolvers.js"), exports);